requirements.txt
Django==5.0.1
djangorestframework==3.14.0
pysnmp>=4.4.12
django-cors-headers==4.3.1
python-dotenv==1.0.0
